+++ src/browser/get_document.rs
@@ -1,0 +1,11 @@
+use super::*;
+
+pub(crate) trait GetDocument {
+  fn get_document(&self) -> Document;
+}
+
+impl GetDocument for Window {
+  fn get_document(&self) -> Document {
+    self.document().expect("`window.document` missing")
+  }
+}
