+++ src/browser/window.rs
@@ -1,0 +1,5 @@
+use super::*;
+
+pub(crate) fn window() -> Window {
+  web_sys::window().expect("`window` missing")
+}
